web/requirements.txt
flask
richcolorlog
toml
paho-mqtt